Run migrations before deploying new image
In most cases, deploying a new image before running migrations will
break the app, because it's referencing models that haven't been updated
at the DB level.

This changes the migrate task to execute migrations against the running
compose environment without pulling and restarting the portal service.
The workflow is now:

  make migrate
  make deploy service=portal