Make portal-deploy mounts old volume, masking host folder.
On restart of a container, the previous volume is left mounted. The
volume must be removed in order to apply changes from the mounted host folder.

Therefore, the step `docker-compose rm --stop portal` has beed added to the
deploy target.

This exposes a number of migrations that have not been properly added.