Use full chain for dev LE certs
Nginx was configured to only serve the leaf certificate without any intermediates. It's unclear how browsers work around this, presumably this is a common enough misconfiguration that they have some special handling for it. Most other tooling like cURL and OpenSSL will fail the cert validation step b/c the local issuer is not returned as part of the cert chain, making the leaf cert impossible to validate in isolation.