Bump portal version to latest
The new GH actions builder only builds latest. Building on the Mac
M1 chip is building wrong, even when setting the platform to amd64. In
an effort to deploy for this changelog, this will deploy latest. Once we
have more time we can revert to using individual tags per release.